third_party/WebKit/LayoutTests/platform/android/inspector/layer-compositing-reasons-expected.txt

translated

backface hidden
animated
opacity
reflected
perspective
preserve3d
Compositing reasons for #document: root
Compositing reasons for div#transform3d: assumedOverlap,transform3D
Compositing reasons for iframe#iframe: iFrame
Compositing reasons for div#backface-visibility: assumedOverlap,backfaceVisibilityHidden
Compositing reasons for div#animation: animation,assumedOverlap
Compositing reasons for div#transformWithCompositedDescendants: assumedOverlap,transformWithCompositedDescendants
Compositing reasons for div#opacityWithCompositedDescendants: assumedOverlap,opacityWithCompositedDescendants
Compositing reasons for div#reflectionWithCompositedDescendants: reflectionWithCompositedDescendants
Compositing reasons for div#perspective: assumedOverlap,perspective
Compositing reasons for div#preserve3d: assumedOverlap,preserve3D
